+++ src/main/java/net/archers/config/TweaksConfig.java
@@ -5,5 +5,5 @@ public class TweaksConfig { public TweaksConfig() {}
 //    public float quick_charge_enchantment_multiplier_per_level = 0.1F;
 //    public float serverside_crossbow_charging_tolerance = 0.95F;
     public boolean enable_infinity_for_crossbows = true;
-    public float hunters_mark_damage_per_stack = 0.05F;
+    public float hunters_mark_damage_per_stack = 0.6F;
 }