+++ src/main/java/net/archers/effect/Effects.java
@@ -10,6 +10,7 @@ import net.minecraft.registry.Registry;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
 import net.spell_engine.api.effect.HealthImpacting;
+import net.spell_engine.api.effect.SpellStash;
 import net.spell_engine.api.effect.Synchronized;
 
 import java.util.ArrayList;
@@ -36,9 +37,10 @@ public class Effects {
         }
     }
 
+    public static final Entry MARKER_SHOT = new Entry("marker_shot",
+            new CustomStatusEffect(StatusEffectCategory.BENEFICIAL, 0xff0000));
     public static final Entry HUNTERS_MARK = new Entry("hunters_mark", 
             new CustomStatusEffect(StatusEffectCategory.HARMFUL, 0xff0000));
-    
     public static final Entry ENTANGLING_ROOTS = new Entry("entangling_roots",
             new CustomStatusEffect(StatusEffectCategory.HARMFUL, 0x993333));
 
@@ -50,7 +52,7 @@ public class Effects {
                 ENTANGLING_ROOTS.modifierId(),
                 -0.5F,
                 EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE);
-        
+        SpellStash.configure(MARKER_SHOT.effect, Identifier.of(ArchersMod.ID, "power_shot"), 1);
         Synchronized.configure(HUNTERS_MARK.effect, true);
         HealthImpacting.configureDamageTaken(HUNTERS_MARK.effect, config.hunters_mark_damage_per_stack);
         Synchronized.configure(ENTANGLING_ROOTS.effect, true);